fixed restype setting

+++ C/so_py_tester.py
@@ -12,9 +12,9 @@ def main():
     
     # tell python the function will return a void pointer
     cs_lib.makeMdaStruct.restype = c_void_p
-    cs_lib.calculateChi.restype = c_float
-    cs_lib.calculateLnLiklihood.restype = c_float
-    cs_lib.calculateLnLiklihoodResids.restype = c_float
+    cs_lib.calculateChi.restype = c_double
+    cs_lib.calculateLnLiklihood.restype = c_double
+    cs_lib.calculateLnLiklihoodResids.restype = c_double
     
     # make the strcuture to hold the MdaData with 50 data points and 8 L dists
     t1 = time.time()
@@ -23,7 +23,7 @@ def main():
     print "make struct, took", int((t2-t1)*1000000), "microseconds"
     
     # make an array with the test data
-    divDat = np.array([0.0, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9], np.float32)
+    divDat = np.array([0.0, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9], np.float64)
     ddPtr = divDat.ctypes.data
     
     # run the function to load the array into the struct
@@ -41,7 +41,7 @@ def main():
     print "set dist x3, took", int((t2-t1)*1000000), "microseconds"
     
     # make an array with the test data
-    params = np.array([0.2, 0.2, 0.2], np.float32)
+    params = np.array([0.2, 0.2, 0.2], np.float64)
     parPtr = params.ctypes.data
     
     # run the function to calculate a chi^2
@@ -63,7 +63,7 @@ def main():
     print "  the log liklihood was:",lnlik
     
     # run the function to calculate a lnliklihood with external resids
-    resids = np.arange(10.0, dtype=np.float32)
+    resids = np.arange(10.0, dtype=np.float64)
     resPtr = resids.ctypes.data
     t1 = time.time()
     lnlik = cs_lib.calculateLnLiklihoodResids(mdaData, parPtr, resPtr)
